refactoring: separate mainloop (and thread), World becomes mainly a facade for the Clojure side.
