+++ src/clojure/yaw/world.clj
@@ -165,7 +165,7 @@
 
 (defn camera "Retrieve the main camera of the world" [world] (.getCamera world))
 
-(defn clear-cameras! "Remove all the cameras from the `world`" [world] (.emptyListCamera world))
+(defn clear-cameras! "Remove all the cameras from the `world`" [world] (.clearCameras world))
 
 (defn add-camera!
   "Add a camera to the `world`"
