+++ test/java/yaw/DifferentLights.java
@@ -69,15 +69,8 @@ public class DifferentLights implements UpdateCallback {
 
         //world.registerUpdateCallback(rCube);
 
-        Thread th = new Thread(world);
-        th.start();
-
-
-        try {
-            th.join();
-        } catch (InterruptedException e) {
-            e.printStackTrace();
-        }
+        world.launch();
+        world.waitFortermination();
     }
 
 }
