+++ test/java/yaw/HitboxRotate.java
@@ -128,16 +128,8 @@ public class HitboxRotate implements UpdateCallback {
         world.registerUpdateCallback(mBb);
 
 
-        Thread th = new Thread(world);
-        th.start();
-
-
-
-        try {
-            th.join();
-        } catch (InterruptedException e) {
-            e.printStackTrace();
-        }
+        world.launch();
+        world.waitFortermination();
     }
