+++ test/java/yaw/RotatingCube.java
@@ -66,16 +66,9 @@ public class RotatingCube implements UpdateCallback {
 
 
 		world.registerUpdateCallback(rCube);
-		
-		Thread th = new Thread(world);
-		th.start();
-		
-		
-		try {
-			th.join();
-		} catch (InterruptedException e) {
-			e.printStackTrace();
-		}
+
+		world.launch();
+		world.waitFortermination();
 	}
 
 }
