+++ test/java/yaw/RotatingCubeLight.java
@@ -70,15 +70,9 @@ public class RotatingCubeLight implements UpdateCallback {
 
         world.registerUpdateCallback(rCube);
 
-        Thread th = new Thread(world);
-        th.start();
+        world.launch();
+        world.waitFortermination();
 
-
-        try {
-            th.join();
-        } catch (InterruptedException e) {
-            e.printStackTrace();
-        }
     }
 
 }
