+++ test/java/yaw/RotatingGroup.java
@@ -88,15 +88,8 @@ public class RotatingGroup implements UpdateCallback {
         world.registerUpdateCallback(rGroup);
 
 
-        Thread th = new Thread(world);
-        th.start();
-
-
-        try {
-            th.join();
-        } catch (InterruptedException e) {
-            e.printStackTrace();
-        }
+        world.launch();
+        world.waitFortermination();
     }
 
 }
