+++ test/java/yaw/TestCeilingCamera.java
@@ -68,15 +68,8 @@ public class TestCeilingCamera implements UpdateCallback {
 
         world.registerUpdateCallback(ceilingCamera);
 
-        Thread th = new Thread(world);
-        th.start();
-
-
-        try {
-            th.join();
-        } catch (InterruptedException e) {
-            e.printStackTrace();
-        }
+        world.launch();
+        world.waitFortermination();
     }
 
 }
