+++ test/java/yaw/TestGroundCamera.java
@@ -69,15 +69,9 @@ public class TestGroundCamera implements UpdateCallback {
 
         //world.registerUpdateCallback(rCube);
 
-        Thread th = new Thread(world);
-        th.start();
+        world.launch();
+        world.waitFortermination();
 
-
-        try {
-            th.join();
-        } catch (InterruptedException e) {
-            e.printStackTrace();
-        }
     }
 
 }
