+++ test/java/yaw/TestHitBoxDifferentPlan.java
@@ -148,16 +148,6 @@ public class TestHitBoxDifferentPlan implements UpdateCallback {
         world.registerUpdateCallback(mBb);
 
 
-        Thread th = new Thread(world);
-        th.start();
-
-
-
-        try {
-            th.join();
-        } catch (InterruptedException e) {
-            e.printStackTrace();
-        }
     }
